DEV: use logging for logs

+++ pdfstream/main.py
@@ -6,8 +6,8 @@ import time
 
 import fire
 
+from pdfstream.callbacks.config import Config
 from pdfstream.callbacks.ananlysisserver import AnalysisServer
-from pdfstream.callbacks.config import Config, ConfigError
 from pdfstream.callbacks.serializationserver import SerializationServer
 from pdfstream.callbacks.visualizationserver import VisualizationServer
 
@@ -38,16 +38,15 @@ def main():
     return
 
 
-def create_logger(log_file: str):
+def create_logger(log_file: T.Union[str, Path]) -> None:
     logger = multiprocessing.get_logger()
     logger.setLevel(logging.INFO)
-    formatter = logging.Formatter(\
-        '[%(asctime)s| %(levelname)s| %(processName)s] %(message)s')
-    handler = logging.FileHandler(log_file)
-    handler.setFormatter(formatter)
     # this bit will make sure you won't have 
     # duplicated messages in the output
     if not len(logger.handlers): 
+        formatter = logging.Formatter('[%(asctime)s | %(levelname)s | %(processName)s] %(message)s')
+        handler = logging.FileHandler(log_file)
+        handler.setFormatter(formatter)
         logger.addHandler(handler)
     return logger
 
@@ -55,7 +54,12 @@ def create_logger(log_file: str):
 def _start_server(server: T.Any, cfg_file: str) -> None:
     config = Config()
     config.read_a_file(cfg_file)
-    return server(config).start()
+    server = server(config)
+    try:
+        server.start()
+    except KeyboardInterrupt:
+        server.stop()
+    return
 
 
 def _run_server(cfg_file: str) -> None:
@@ -67,16 +71,13 @@ def _run_server(cfg_file: str) -> None:
     cfg_path = Path(cfg_file)
     if not cfg_path.is_file():
         cfg_file = find_cfg_file(CONFIG_DIR, cfg_file)
-    config = Config()
-    config.read_a_file(cfg_file)
-    create_logger(str(config.log_file))
     servers = (AnalysisServer, SerializationServer, VisualizationServer)
-    processes = []
+    processes: T.List[Process] = []
     for server in servers:
         p = Process(
             target=_start_server, 
             args=(server, cfg_file), 
-            name="pdfstream"
+            name="pdfstream_server"
         )
         processes.append(p)
     for p in processes:
@@ -85,9 +86,13 @@ def _run_server(cfg_file: str) -> None:
         while True:
             time.sleep(0.5)
     except KeyboardInterrupt:
+        pass
+    finally:
         for p in processes:
-            p.kill()
+            if p.is_alive:
+                p.kill()
             p.join()
+            p.close()
     return
